Rule of five
This may be needed to implement in the LinkedQueue class. In the
circular list class, the move constructor and assignment operator are
implemented, but the copy methods are explicitly deleted. This could'nt
be implemented because the rule of five is not implemented at the
LinkedQueue, as a CircularList of cashiers(who has a linked queue) is
instantiated.

Another solution may be changing the LinkedQueue to an ArrayQueue.